% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVHBA_SVM.R
\name{LVHBA_SVM}
\alias{LVHBA_SVM}
\title{Solving support vector machine hyper-parameter selection problem based on LV-HBA}
\usage{
LVHBA_SVM(
  features,
  labels,
  K = 3,
  N = 200,
  alpha = 0.002,
  beta = 0.002,
  eta = 0.003,
  gamma1 = 1,
  gamma2 = 1,
  c_0 = 2,
  power = 0.48,
  r = 10,
  ub = 10,
  lb = 10^(-6),
  tol = 0.001,
  auto_tuning = FALSE,
  temperature = 0.1
)
}
\arguments{
\item{features}{The feature matrix with dimensions n by p, where the j-th column corresponds to the j-th feature.}

\item{labels}{The label vector where the j-th element represents the label of the j-th feature,
with a value of 1 indicating the positive class and -1 indicating the negative class.}

\item{K}{K-fold cross validation. Defult is 3.}

\item{N}{Total iterations. Default is 200.}

\item{alpha}{Projection stepsize in LV-HBA of \code{x,y} which is fixed. Default is 0.002.}

\item{beta}{Projection stepsize in LV-HBA of \code{z} which is fixed. Default is 0.002.}

\item{eta}{Projection stepsize  in LV-HBA of \code{theta, lambda} which is fixed. Default is 0.003.}

\item{gamma1}{Proximal parameter in LV-HBA. Default is 1.}

\item{gamma2}{Proximal parameter in LV-HBA. Default is 1.}

\item{c_0}{which is used to generate the penalty parameter \eqn{c_k} in LV-HBA by \eqn{c_k = \underline{c}(k+1)^{power}} where \eqn{\underline{c} = c_0}.
Default is 2.}

\item{power}{The power exponent in \eqn{c_k = \underline{c}(k+1)^power}. Default is 0.48.}

\item{r}{parameter in LV-HBA. Default is 10.}

\item{ub}{\eqn{\bar{w}_{ub}=ub*\mathbf{1}}, where \eqn{\mathbf{1}} is a vector of ones with the same size as  \eqn{\bar{w}}. Default is 10.}

\item{lb}{\eqn{\bar{w}_{lb}=lb*\mathbf{1}}, where \eqn{\mathbf{1}} is a vector of ones with the same size as  \eqn{\bar{w}}. Default is \eqn{10^{-6}}.}

\item{tol}{Tolerance. IF \eqn{|(x^{k + 1}, y^{k + 1}) - (x^{k}, y^k)|/\sqrt{1+|(x^{k}, y^k)|^2} < tol},
then terminate the iteration, where \eqn{x^k} represents the upper-level variable, \eqn{y^k} represents the lower-level variable. Default is 0.001.}

\item{auto_tuning}{Whether an auto-hyperparameter-tuning is needed. Default is
\code{FALSE}.}

\item{temperature}{Temperature of simulating annealing method for auto-
hyper-parameter-tuning. Default is 0.1.}
}
\value{
\item{x}{The optimal hyper-parameters, which is a vector of length (p+1).
The first value is regularization parameter, and the remaining element stands for box constraint \eqn{\bar{w}}}
\item{y}{A vector combines by variables and auxiliary variables of the subproblems.}
\item{w}{A list of t vectors.
The t-th element of the list \code{w} (which is a vector),
and the t-th value of c (which is a scalar) represents the variables of the t-th subproblem.}
\item{c}{A vector of t elements.
The t-th element of the list \code{w} (which is a vector), and the t-th value of c (which is a scalar) represents the variables of the t-th subproblem.}
\item{xi}{Auxiliary variable}
\item{X_seq}{A list. Describe the sequence x generated by MEHA in the iterative process.}
\item{Y_seq}{A list. Describe the sequence y generated by MEHA in the iterative process.}
\item{F_seq}{The upper function value sequence generated in the iterative process.}
\item{val_error}{The upper function value in the last iteration.}
}
\description{
This R function is written to solve the hyper-parameter selection problem of Support Vector Machines (SVM) using the LV-HBA,
please refer to the listed literature for the specific algorithm and model. It should be noted that in order to successfully apply the algorithm,
the upper objective function is smoothed using softplus function approximation,
and the lower problem is smoothed by introducing auxiliary variables and additional constraints.
}
\references{
Yao, W., Yu, C., Zeng, S., & Zhang, J. (2024).
"Constrained Bi-Level Optimization: Proximal Lagrangian Value Function Approach and Hessian-free Algorithm."
Available at: https://openreview.net/forum?id=xJ5N8qrEPl

\if{html}{\out{<div class="sourceCode">}}\preformatted{Gao, L., Ye, J. J., Yin, H., Zeng, S., & Zhang, J. (2022).
"Value function based difference-of-convex algorithm for bilevel hyperparameter selection problems."
Available at: https://proceedings.mlr.press/v162/gao22j.html
}\if{html}{\out{</div>}}
}
